use a more sensible default for selector option
As a [data-depth] attribute is required for each layer to move, it makes sense for the attribute to be the default selector for layers. With this configuration, there should be no need to define a selector - just add the attribute to the elements you want to apply parallax to, and it just works.